+++ CHANGELOG.md
@@ -45,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - sdk(programs): Added `gmsol-competition` support under `competition` feature.
 - sdk(sdk): Added support for `gmsol-competition` via `CompetitionOps`, under `competition` feature.
 - sdk(sdk): Added `StoreOps::initialize_callback_authority` function.
+- sdk(sdk): Added `KeyedAccount`, implementing `gmsol_decode::AccountAccess`.
 
 ### Fixed